requirements.txt
bleach==6.2.0
certifi==2025.6.15
psycopg2-binary~=2.9.10
chardet==5.2.0
charset-normalizer==3.4.2
click~=8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask~=3.1.2
fonttools==4.58.4
gunicorn==23.0.0
idna==3.10
itsdangerous~=2.2.0
Jinja2~=3.1.6
joblib==1.3.2              # changed to match sklearn 1.3.2
kaggle==1.7.4.5
kiwisolver==1.4.8
MarkupSafe~=3.0.2
matplotlib==3.7.5
numpy~=2.3.2
packaging==25.0
pandas~=2.3.2
pickle-mixin==1.0.2
pillow==11.2.1
protobuf==6.31.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.4
scikit-learn==1.3.2
scipy==1.12.0
six==1.17.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug~=3.1.3

SQLAlchemy~=2.0.43
pip~=25.2
typing_extensions~=4.14.1
blinker~=1.9.0
WTForms~=3.2.1
greenlet~=3.2.4